+++ src/shared/validation-rules.js
@@ -47,6 +47,8 @@ export const alphaNumSpaceDotCommaHyphen = (value) => {
   return /^[0-9a-zA-Z #.,-]*$/.test(value)
 };
 
+// 0-9a-zA-Z #'.,-
+// !@$%^&*(){}[]:;"<>?\|*+
 export const alphaNumSpaceDotCommaHyphenApostrophe = (value) => {
   if (typeof value === 'undefined' || value === null || value === '') {
     return true
@@ -54,6 +56,8 @@ export const alphaNumSpaceDotCommaHyphenApostrophe = (value) => {
   return /^[0-9a-zA-Z #'.,-]*$/.test(value)
 };
 
+// 0-9a-zA-Z /,.#-
+// !@$%^&*(){}[]:;"<>?\|*+
 export const address = (value) => {
   if (typeof value === 'undefined' || value === null || value === '') {
     return true
@@ -61,11 +65,29 @@ export const address = (value) => {
   return /^[0-9a-zA-Z /,.#-]*$/.test(value)
 };
 
+// a-zA-Z -'
+// !@#$%^&*(){}[]:;"<>?/\|*+1
 export const city = (value) => {
   if (typeof value === 'undefined' || value === null || value === '') {
     return true
   }
-  return /^[a-zA-Z -]*$/.test(value)
+  return /^[a-zA-Z '-]*$/.test(value)
+};
+
+// !@#$%^&*(){}[]:;"<>?/\|*+1,.
+export const name = (value) => {
+  if (typeof value === 'undefined' || value === null || value === '') {
+    return true
+  }
+  return /^[a-zA-Z '-]*$/.test(value)
+};
+
+// !@#$%^&*(){}[]:;"<>?/\|*+1
+export const namePrefix = (value) => {
+  if (typeof value === 'undefined' || value === null || value === '') {
+    return true
+  }
+  return /^[a-zA-Z /,.'-]*$/.test(value)
 };
 
 export const companyName = (value) => {
@@ -75,9 +97,25 @@ export const companyName = (value) => {
   return /^[a-zA-Z #'.,&-]*$/.test(value)
 };
 
+// 01 )(-
+// asAS!@#$%^&*{}[]:;"<>?/\|*+
 export const usaPhone = (value) => {
   if (typeof value === 'undefined' || value === null || value === '') {
     return true
   }
   return /^[0-9 )(-]*$/.test(value)
+};
+
+export const usaZip = (value) => {
+  if (typeof value === 'undefined' || value === null || value === '') {
+    return true
+  }
+  return /^[0-9]{5}(?:-[0-9]{4})?$/.test(value)
+};
+
+export const myUrl = (value) => {
+  if (typeof value === 'undefined' || value === null || value === '') {
+    return true
+  }
+  return /^(http|https):\/\/[^ "]+$/.test(value)
 };